Delay processing for non-internal edges.
In our union-find, we merge two trees by attaching the tree with
higher root value to the tree with lower one. In addition, the
of a node in union-find cannot be changed directly after it is
inserted. Therefore, external vertices should be inserted after
internal vertices. Otherwise, an unprocess internal vertices can
be insert into the union-find with the wrong value.

Example:
Consider two vertices with their respective adjacency list.

2, (2, 1), (2, 3)
1, (1, 2), (1, 3)

The sequence of operation is
- Buffer edge (2, 1), (2, 3) since we don't know if 1 or 3 are
internal or not.
- Process edge (1, 2), create union between 1 and 2,
since we know 1 and 2 are internal vertices.
- Buffer edge(1, 3), since we don't know if 3 is internal or not.

After the entire partition is processed, create union for all
buffered edge 2 and 1, 2 and 3, 1 and 3